insightface/face_recognition/face_detection/requirements.txt
absl-py==0.8.1
astor==0.8.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cmake==3.18.2.post1
dlib==19.18.0
face-recognition==1.2.3
face-recognition-models==0.3.0
filelock==3.0.12
future==0.18.2
gast==0.2.2
gdown==3.12.2
google-pasta==0.1.7
greenlet==0.4.17
grpcio==1.24.1
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
imutils==0.5.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mock==3.0.5
msgpack==1.0.0
mtcnn==0.0.9
neovim==0.3.1
numpy==1.17.3
onnx==1.6.0
onnx-tf==1.3.0
onnxruntime==0.5.0
opencv-python==4.1.1.26
opt-einsum==3.1.0
Pillow==6.2.1
protobuf==3.10.0
pynvim==0.4.2
PySocks==1.7.1
PyYAML==5.1.2
requests==2.24.0
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.50.2
typing-extensions==3.7.4
urllib3==1.25.10
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.2
zipp==3.1.0
